+++ examples/main/modules.tf
@@ -1,0 +1,108 @@
+module "azure_region" {
+  source  = "claranet/regions/azurerm"
+  version = "x.x.x"
+
+  azure_region = var.azure_region
+}
+
+module "rg" {
+  source  = "claranet/rg/azurerm"
+  version = "x.x.x"
+
+  location    = module.azure_region.location
+  client_name = var.client_name
+  environment = var.environment
+  stack       = var.stack
+}
+
+module "logs" {
+  source  = "claranet/run-common/azurerm//modules/logs"
+  version = "x.x.x"
+
+  client_name    = var.client_name
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  environment    = var.environment
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+}
+
+
+module "azure_network_vnet" {
+  source  = "claranet/vnet/azurerm"
+  version = "x.x.x"
+
+  environment    = var.environment
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+  vnet_cidr           = ["10.10.0.0/16"]
+}
+
+module "azure_network_subnet" {
+  source  = "claranet/subnet/azurerm"
+  version = "x.x.x"
+
+  environment    = var.environment
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+
+  virtual_network_name = module.azure_network_vnet.virtual_network_name
+  subnet_cidr_list     = ["10.10.10.0/24"]
+}
+
+module "vm" {
+  source  = "claranet/linux-vm/azurerm"
+  version = "x.x.x"
+
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  client_name    = var.client_name
+  environment    = var.environment
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+
+  subnet_id      = module.azure_network_subnet.subnet_id
+  vm_size        = "Standard_B2s"
+  custom_name    = "app-${var.stack}-${var.client_name}-${module.azure_region.location_short}-${var.environment}-vm"
+  admin_username = var.vm_administrator_login
+  ssh_public_key = var.ssh_public_key
+
+  zone_id = 1
+
+  vm_image = {
+    publisher = "Debian"
+    offer     = "debian-10"
+    sku       = "10"
+    version   = "latest"
+  }
+
+  diagnostics_storage_account_name      = module.logs.logs_storage_account_name
+  diagnostics_storage_account_sas_token = module.logs.logs_storage_account_sas_token
+}
+
+module "vm_logs" {
+  source  = "claranet/vm-logs/azurerm"
+  version = "x.x.x"
+
+  environment = var.environment
+  stack       = var.stack
+
+  vm_id = module.vm.vm_id
+
+  diagnostics_storage_account_name      = module.logs.logs_storage_account_name
+  diagnostics_storage_account_sas_token = module.logs.logs_storage_account_sas_token
+
+  tags = {
+    environment = var.environment
+    stack       = var.stack
+  }
+}