+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Added
   * AZ-615: Add an option to enable or disable default tags
 
+Changed
+  * AZ-572: Revamp examples and improve CI
+
 # v4.0.1 - 2021-08-27
 
 Updated